subscriptions_view: Fully functional implementation
User can add, remove and edit clients, see asynchronously connection
state and select a client to see their subscriptions.

In the subscriptions overview panel they can see the subscriptions for
the selected client, if it's either subscribed or not, remove them and
edit them.

In subscriptions overview panel's top part there are some buttons to
perform various actions that may be of interest to the user, there is a
button to show the incoming messages for the selected client, when
pressed it will present a bottom sheet with a list of messages and
various metadata for each message.

I'm proud of this commit, peace <3.